+++ src/features/card/pages/add-card/index.tsx
@@ -1,10 +1,10 @@
 import { CardInputContainer } from '@/features/card/components/CardInputContainer';
 import { IconArrowBack } from '@/assets/IconArrowBack';
-import { Header } from '@/shared/components/molecules/Header';
-import { Button } from '@/shared/components/atoms/Button';
-import { Heading } from '@/shared/components/atoms/Heading';
-import { RootLayout } from '@/shared/layout/RootLayout';
-import { AppLayout } from '@/shared/layout/AppLayout';
+import { Header } from '@/components/molecules/Header';
+import { Button } from '@/components/atoms/Button';
+import { Heading } from '@/components/atoms/Heading';
+import { RootLayout } from '@/components/layout/RootLayout';
+import { AppLayout } from '@/components/layout/AppLayout';
 
 interface Props {
   onPrev: () => void;
